+++ Sm5sh/Mods/Sm5sh.Mods.Music/Models/BgmEntry.cs
@@ -103,13 +103,13 @@ namespace Sm5sh.Mods.Music.Models
         {
             public BgmEntry Parent { get; }
             public string NameId { get { return Parent.ToneId; } }
-            public ulong LoopStartMs { get; set; }
-            public ulong LoopStartSample { get; set; }
-            public ulong LoopEndMs { get; set; }
-            public ulong LoopEndSample { get; set; }
-            public ulong TotalTimeMs { get; set; }
-            public ulong TotalSamples { get; set; }
-            public ulong Frequency { get { return TotalTimeMs / 1000 * TotalSamples; } }
+            public uint LoopStartMs { get; set; }
+            public uint LoopStartSample { get; set; }
+            public uint LoopEndMs { get; set; }
+            public uint LoopEndSample { get; set; }
+            public uint TotalTimeMs { get; set; }
+            public uint TotalSamples { get; set; }
+            public uint Frequency { get { return TotalTimeMs / 1000 * TotalSamples; } }
 
             public BgmPropertyEntry(BgmEntry parent)
             {
